+++ router/core/graph_server.go
@@ -367,7 +367,9 @@ func newGraphServer(routerCtx context.Context, r *Router, response *routerconfig
 			LivenessCheckPath:   s.livenessCheckPath,
 			CompositePropagator: s.compositePropagator,
 			TracerProvider:      s.tracerProvider,
-			SpanNameFormatter:   SpanNameFormatter,
+			SpanNameFormatter: func(_ string, r *http.Request) string {
+				return s.spanNameFormatter(r)
+			},
 		})
 		httpRouter.Use(handler)
 	}
@@ -1363,10 +1365,12 @@ func (s *graphServer) buildGraphMux(
 		logger:           s.logger,
 		trackUsageInfo:   s.graphqlMetricsConfig.Enabled || s.metricConfig.Prometheus.PromSchemaFieldUsage.Enabled,
 		subscriptionClientOptions: &SubscriptionClientOptions{
-			PingInterval: s.engineExecutionConfiguration.WebSocketClientPingInterval,
-			PingTimeout:  s.engineExecutionConfiguration.WebSocketClientPingTimeout,
-			ReadTimeout:  s.engineExecutionConfiguration.WebSocketClientReadTimeout,
-			FrameTimeout: s.engineExecutionConfiguration.WebSocketClientFrameTimeout,
+			PingInterval:              s.engineExecutionConfiguration.WebSocketClientPingInterval,
+			PingTimeout:               s.engineExecutionConfiguration.WebSocketClientPingTimeout,
+			WriteTimeout:              s.engineExecutionConfiguration.WebSocketClientWriteTimeout,
+			AckTimeout:                s.engineExecutionConfiguration.WebSocketClientAckTimeout,
+			ReadLimit:                 int64(s.engineExecutionConfiguration.WebSocketClientReadLimit),
+			DefaultErrorExtensionCode: s.subgraphErrorPropagation.DefaultExtensionCode,
 		},
 		transportOptions: &TransportOptions{
 			SubgraphTransportOptions:      s.subgraphTransportOptions,
@@ -1377,6 +1381,7 @@ func (s *graphServer) buildGraphMux(
 			RetryOptions:                  *processedRetryOptions,
 			TracerProvider:                s.tracerProvider,
 			TracePropagators:              s.compositePropagator,
+			SpanNameFormatter:             s.spanNameFormatter,
 			LocalhostFallbackInsideDocker: s.localhostFallbackInsideDocker,
 			Logger:                        s.logger,
 			EnableTraceClient:             enableTraceClient,
@@ -1725,6 +1730,7 @@ func (s *graphServer) buildGraphMux(
 		HasPreOriginHandlers:                   len(s.preOriginHandlers) != 0,
 		HeaderPropagation:                      s.headerPropagation,
 		OperationContentAttributes:             s.traceConfig.OperationContentAttributes,
+		SpanNameFormatter:                      s.spanNameFormatter,
 	})
 
 	if s.webSocketConfiguration != nil && s.webSocketConfiguration.Enabled {
@@ -1738,11 +1744,11 @@ func (s *graphServer) buildGraphMux(
 			AccessController:          s.accessController,
 			Logger:                    s.logger,
 			Stats:                     s.engineStats,
-			ReadTimeout:               s.engineExecutionConfiguration.WebSocketClientReadTimeout,
-			WriteTimeout:              s.engineExecutionConfiguration.WebSocketClientWriteTimeout,
+			ReadTimeout:               s.engineExecutionConfiguration.WebSocketServerReadTimeout,
+			WriteTimeout:              s.engineExecutionConfiguration.WebSocketServerWriteTimeout,
 			EnableNetPoll:             s.engineExecutionConfiguration.EnableNetPoll,
-			NetPollTimeout:            s.engineExecutionConfiguration.WebSocketClientPollTimeout,
-			NetPollConnBufferSize:     s.engineExecutionConfiguration.WebSocketClientConnBufferSize,
+			NetPollTimeout:            s.engineExecutionConfiguration.WebSocketServerPollTimeout,
+			NetPollConnBufferSize:     s.engineExecutionConfiguration.WebSocketServerConnBufferSize,
 			WebSocketConfiguration:    s.webSocketConfiguration,
 			ClientHeader:              s.clientHeader,
 			DisableVariablesRemapping: s.engineExecutionConfiguration.DisableVariablesRemapping,
@@ -1831,7 +1837,6 @@ func (s *graphServer) setupConnector(
 				Logger:   s.logger,
 				Endpoint: sg.RoutingUrl,
 			})
-
 			if err != nil {
 				return fmt.Errorf("failed to create standalone plugin for subgraph %s: %w", dsConfig.Id, err)
 			}
@@ -1861,6 +1866,7 @@ func (s *graphServer) setupConnector(
 		getTraceAttributes := CreateGRPCTraceGetter(
 			telemetryAttributeExpressions,
 			tracingAttributeExpressions,
+			s.spanNameFormatter,
 		)
 
 		if imgRef := pluginConfig.GetImageReference(); imgRef != nil {